Changed path to data

+++ precipitation_forecasting/rainyday_labeler.py
@@ -7,10 +7,12 @@ from tqdm import tqdm
 import h5py
 import config
 
-radar_dir =  config.dir_rtcor 
+radar_dir =  config.dir_rtcor
 label_dir = config.dir_labels
 
-files = sorted([f for f in listdir(radar_dir) if isfile(join(radar_dir, f)) and f[16:20] == '2019' and f.endswith('.h5')])
+#files = sorted([f for f in listdir(radar_dir) if isfile(join(radar_dir, f)) and f[16:20] == '2019' and f.endswith('.h5')])
+root = radar_dir + '2019'
+files = sorted([name for path, subdirs, files in os.walk(root) for name in files])
 
 cluttermask = ~np.load('cluttermask.npy')
 def is_rainy(rdr):    
@@ -56,15 +58,23 @@ def make_dir(dir_name):
   '''
   if not os.path.exists(dir_name):
     os.makedirs(dir_name)
-    
+
 make_dir(label_dir)
 for f in tqdm(files):
-    try:
-        rdr = load_h5(radar_dir+f)
-        rainy = is_rainy(rdr)
-    except Exception as e:
-        print(e)
-        rainy = False
-        
-    label_fn = label_dir + f
-    np.save(label_fn, rainy) 
+    ts = f.replace('RAD_NL25_RAC_RT_', '')
+    ts = ts.replace('.h5', '')
+    label_fn = label_dir +  ts + '.npy'
+
+    
+    if not os.path.isfile(label_fn):
+        count+=1
+        try:
+            year = ts[:4]
+            month = ts[4:6]
+            rdr = load_h5(radar_dir+'/{}/{}/{}'.format(year,month,f))
+            rainy = is_rainy(rdr)
+        except Exception as e:
+            rainy = False
+            print(e)
+        np.save(label_fn, rainy)
+
